long id changes

+++ ubbprotool-pc/src/main/java/com/chk/ubbprotool/ubbprotool/Controller/UniversityClassController.java
@@ -60,7 +60,7 @@ public class UniversityClassController {
     }
 
     @PostMapping("/getPosiibleClassesForChanged")
-    public ResponseEntity<List<UniversityClassDTO>> getPossibleClassesForChange(@RequestParam("classId") int classId , @RequestParam("date")String curentDate)
+    public ResponseEntity<List<UniversityClassDTO>> getPossibleClassesForChange(@RequestParam("classId") Long classId , @RequestParam("date")String curentDate)
     {
         Date currentDate=Date.valueOf(curentDate);//converting string into sql date
         return ResponseEntity.ok(universityClassService.getPossibleClassesToBeChanged(classId, currentDate));
